api/requirements.txt
# langchain-google-genai
langchain_google_genai
# langchain_core
langchain-core
langchain_community
langserve
langchain_ollama
python-dotenv
streamlit
# langchain
fastapi
uvicorn
sse_starlette
